order_app/requirements.txt
alembic==1.7.7
attrs==21.4.0
blinker==1.4
boto3==1.23.3
botocore==1.26.3
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
Mako==1.2.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pycryptodome==3.14.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.36
tabulate==0.8.9
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
